+++ src/distribution/Histogram.cpp
@@ -5,7 +5,7 @@ Histogram::Histogram(const std::vector<unsigned char>& imageData, int size) : Di
     std::fill(data.begin(), data.end(), 0);
     totalCount = 0;
     // 计算直方图，假设为灰度图像(不分通道)
-    for (int i = 0; i < imageData.size(); ++i) {
+    for (size_t i = 0; i < imageData.size(); ++i) {
         data[imageData[i]]++;
         totalCount++;
     }
